Move serializaiton/deserialization of Registration into another class
This prevents having to add GSON as a dependency in go.jar/lib
